+++ CHANGELOG.md
@@ -10,5 +10,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Servers created with Fabric loader >= 0.12.0 not launching [#513]
+- Failure to launch Minecraft causing the main frame to not show itself again
 
 ### Misc