+++ CHANGELOG.md
@@ -18,3 +18,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Misc
 - Remove dependency on pack200 [#516]
+- Update dependencies and gradle wrapper