+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.8.1
+## 3.4.8.2
 
 ### New Features
 
 ### Fixes
-- Default instance title format setting being invalid
-- Error reporting throwing a NPE when message was null
 
 ### Misc